+++ src/propylean/equipments/__init__.py
@@ -2,5 +2,5 @@ from propylean.equipments.exchangers import AirCooler, ElectricHeater
 from propylean.equipments.rotary import CentrifugalPump, PositiveDisplacementPump,\
     CentrifugalCompressor
 from propylean.equipments.static import PipeSegment
-from propylean.equipments.storages import VerticalStorage, StorageBullet, Tank,\
+from propylean.equipments.storages import VerticalStorage, Bullet, Tank, Sphere,\
     AirReciever
