+++ tests/pytest.ini
@@ -1,0 +1,13 @@
+[pytest]
+markers =
+    positive: marks tests as positive.
+    negative: marks tests as negative.
+    instantiation: marks tests as instantiation of an object.
+    unit_change:  marks tests as unit change tests.
+    vol_flowrate
+    viscosity
+    density
+    subtraction
+    addition
+    segment_frame
+    temp
